Bugfixes: - corrected minor bug in greens function
Minor improvements in basis_set and separation matrix. mainly for resilience

Implemented a two-sided Rayleigh quotient iterative method to solve the nonlinear eigenvalue problem.
It has been tested to solve a linear eigenvalue problem for a single eigenvalue
I will add the missing feature to this class. This commit is a checkpoint